breaking it up so the generation of reaction_df and pi_df are in different files
